feat: extract tenant db context and query filters for reusability. Add ITenantDbContext interface.
