+++ src/Endatix.Infrastructure/Data/Abstractions/ITenantDbContext.cs
@@ -1,0 +1,6 @@
+namespace Endatix.Infrastructure.Data.Abstractions;
+
+public interface ITenantDbContext
+{
+    long GetTenantId();
+}
